+++ tests/Expect/Time/toBeWeekend.php
@@ -8,7 +8,7 @@ test('pass', function () {
 
 test('fails', function () {
     expect('2021-10-22')->toBeWeekend();
-})->throws(ExpectationFailedException::class, 'Failed to assert that [2021-10-22 00:00:00] is saturday or sunday');
+})->throws(ExpectationFailedException::class, 'Failed to assert that [2021-10-22 00:00:00] is Saturday or Sunday');
 
 test('pass negated', function () {
     expect('2021-10-22')->not->toBeWeekend();